Web app/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
edge-tts>=6.1.0
python-dotenv>=1.0.0
google-genai>=1.0.0
